Add a way to display the resulting network
Reworked the types completely to make it possible to display the
resulting network's topology at the end of construction.

Comments and such need to be added.

In the future, before the prototype is presented, it is vital to update
the names of function to make them more consistent and easier to
navigate, as at present we have similar
names (createNetwork/makeNetwork/runNetwork)
that might introduce confusion.