Restructure the project and start commenting
Added cleaner module separation, with explicitly defined
exports. Modules Types and Networks are also (almost) fully commented
now.
Gui has been made a module, with Main.hs re-written to simply use it to
show the GUI. In the future, Gui and Shell main files can be created,
with the project serving as a library behind them.